requirements.txt
flask
tensorflow==2.10.0
numpy
Pillow
gunicorn
